Changed to new format and moved all files to old.
