+++ Old/Other_Scripts/anti_vowel.py
