+++ Old/Other_Scripts/factorial.py
