+++ Old/Other_Scripts/is_prime.py
