+++ Old/cal_score_avg.py
