+++ Old/display_point_maps.py
